+++ puma/tests/test_pie_chart.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 """Unit test script for the functions in histogram.py."""
+from __future__ import annotations
 
 import os
 import tempfile