+++ puma/tests/test_var_vs_eff.py
@@ -2,12 +2,14 @@
 
 
 """Unit test script for the functions in var_vs_eff.py."""
+from __future__ import annotations
 
 import os
 import tempfile
 import unittest
 
 import numpy as np
+from ftag import Flavours
 from matplotlib.testing.compare import compare_images
 
 from puma import VarVsEff, VarVsEffPlot
@@ -42,7 +44,7 @@ class VarVsEffTestCase(unittest.TestCase):
             VarVsEff(
                 np.ones(6),
                 np.ones(6),
-                flat_eff_bin=True,
+                flat_per_bin=True,
                 disc_cut=1.0,
                 working_point=0.77,
             )
@@ -50,7 +52,7 @@ class VarVsEffTestCase(unittest.TestCase):
     def test_var_vs_eff_init_flat_eff_no_wp(self):
         """Test var_vs_eff init."""
         with self.assertRaises(ValueError):
-            VarVsEff(np.ones(6), np.ones(6), flat_eff_bin=True, disc_cut=1.0)
+            VarVsEff(np.ones(6), np.ones(6), flat_per_bin=True, disc_cut=1.0)
 
     def test_var_vs_eff_init_disc_cut_wp(self):
         """Test var_vs_eff init."""
@@ -102,7 +104,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=n_bins,
         )
         np.testing.assert_array_almost_equal(
@@ -118,7 +120,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=n_bins,
         )
         np.testing.assert_array_almost_equal(
@@ -134,7 +136,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=n_bins,
         )
         var_plot_comp = VarVsEff(
@@ -171,7 +173,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=n_bins,
         )
         var_plot.y_var_mean, var_plot.y_var_std = var_plot.get("sig_eff")
@@ -186,7 +188,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=n_bins,
         )
         with self.assertRaises(ValueError):
@@ -200,7 +202,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=1,
         )
         var_plot_comp = VarVsEff(
@@ -209,7 +211,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=2,
         )
         with self.assertRaises(ValueError):
@@ -231,7 +233,7 @@ class VarVsEffTestCase(unittest.TestCase):
             x_var_bkg=self.disc_bkg,
             disc_bkg=self.x_var_bkg,
             working_point=self.working_point,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             bins=1,
         )
         mode_options = ["sig_eff", "bkg_eff", "sig_rej", "bkg_rej"]
@@ -297,7 +299,7 @@ class VarVsEffOutputTestCase(unittest.TestCase):
                 figsize=(9, 6),
             )
 
-    def test_output_plot_flat_eff_bin_bkg_rejection(self):
+    def test_output_plot_flat_per_bin_bkg_rejection(self):
         """Test output plot with flat eff per bin - bkg rejection."""
         # define the curves
         ref_light = VarVsEff(
@@ -308,7 +310,7 @@ class VarVsEffOutputTestCase(unittest.TestCase):
             bins=self.bins,
             working_point=0.5,
             disc_cut=None,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             label="reference model",
         )
         better_light = VarVsEff(
@@ -319,7 +321,7 @@ class VarVsEffOutputTestCase(unittest.TestCase):
             bins=self.bins,
             working_point=0.5,
             disc_cut=None,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             linestyle="dashed",
             label="better model (by construction better for $p_T$ > 110)",
         )
@@ -368,7 +370,7 @@ class VarVsEffOutputTestCase(unittest.TestCase):
             bins=self.bins,
             working_point=0.5,
             disc_cut=None,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             label="reference model",
         )
         better_light = VarVsEff(
@@ -379,7 +381,7 @@ class VarVsEffOutputTestCase(unittest.TestCase):
             bins=self.bins,
             working_point=0.5,
             disc_cut=None,
-            flat_eff_bin=True,
+            flat_per_bin=True,
             label="better model (by construction better for $p_T$ > 110)",
         )
         plot_bkg_rej = VarVsEffPlot(
@@ -417,3 +419,39 @@ class VarVsEffOutputTestCase(unittest.TestCase):
                 tol=1,
             ),
         )
+
+    def test_var_vs_eff_info_str_fixed_eff(self):
+        flat_wp_plot = VarVsEffPlot(
+            mode="bkg_eff",
+        )
+        signal = Flavours["bjets"]
+
+        flat_wp_plot.apply_modified_atlas_second_tag(signal=signal, working_point=0.7)
+
+        expected_tag = "70.0% $b$-jet efficiency"
+        self.assertEqual(flat_wp_plot.atlas_second_tag, expected_tag)
+
+    def test_var_vs_eff_info_str_flat_wp(self):
+        flat_wp_plot = VarVsEffPlot(
+            mode="bkg_eff",
+            atlas_second_tag="test",
+        )
+        signal = Flavours["bjets"]
+
+        flat_wp_plot.apply_modified_atlas_second_tag(
+            signal=signal, working_point=0.7, flat_per_bin=True
+        )
+
+        expected_tag = "test\nFlat 70.0% $b$-jet efficiency per bin"
+        self.assertEqual(flat_wp_plot.atlas_second_tag, expected_tag)
+
+    def test_var_vs_eff_info_str_fixed_disc(self):
+        flat_wp_plot = VarVsEffPlot(
+            mode="bkg_eff",
+        )
+        signal = Flavours["bjets"]
+
+        flat_wp_plot.apply_modified_atlas_second_tag(signal=signal, disc_cut=3.0)
+
+        expected_tag = "$D_{b}$ > 3.0"
+        self.assertEqual(flat_wp_plot.atlas_second_tag, expected_tag)