+++ puma/tests/utils/test_utils.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 """Unit test script for the functions in utils/__init__.py."""
+from __future__ import annotations
 
 import os
 import tempfile